server/docmind/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-docx==1.1.0
pypdf2==3.0.1
httpx==0.26.0
openai==1.12.0
anthropic==0.18.0
aiofiles==23.2.1
